+++ services/horizon/internal/integration/parameters_test.go
@@ -2,10 +2,6 @@
 package integration
 
 import (
-	"bytes"
-	"fmt"
-	"github.com/spf13/cobra"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
@@ -214,54 +210,6 @@ func TestDisablePathFinding(t *testing.T) {
 		assert.Nil(t, test.HorizonIngest().Paths())
 		test.Shutdown()
 	})
-}
-
-func TestIngestionFilteringAlwaysDefaultingToTrue(t *testing.T) {
-	t.Run("ingestion filtering flag set to default value", func(t *testing.T) {
-		test := NewParameterTest(t, map[string]string{})
-		err := test.StartHorizon()
-		assert.NoError(t, err)
-		assert.Equal(t, test.HorizonIngest().Config().EnableIngestionFiltering, true)
-	})
-	t.Run("ingestion filtering flag set to false", func(t *testing.T) {
-		test := NewParameterTest(t, map[string]string{"exp-enable-ingestion-filtering": "false"})
-		err := test.StartHorizon()
-		assert.NoError(t, err)
-		assert.Equal(t, test.HorizonIngest().Config().EnableIngestionFiltering, true)
-	})
-}
-
-func TestHelpOutputForNoIngestionFilteringFlag(t *testing.T) {
-	config, flags := horizon.Flags()
-
-	horizonCmd := &cobra.Command{
-		Use:           "horizon",
-		Short:         "client-facing api server for the Stellar network",
-		SilenceErrors: true,
-		SilenceUsage:  true,
-		Long:          "Client-facing API server for the Stellar network. It acts as the interface between Stellar Core and applications that want to access the Stellar network. It allows you to submit transactions to the network, check the status of accounts, subscribe to event streams and more.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			_, err := horizon.NewAppFromFlags(config, flags)
-			if err != nil {
-				return err
-			}
-			return nil
-		},
-	}
-
-	var writer io.Writer = &bytes.Buffer{}
-	horizonCmd.SetOutput(writer)
-
-	horizonCmd.SetArgs([]string{"-h"})
-	if err := flags.Init(horizonCmd); err != nil {
-		fmt.Println(err)
-	}
-	if err := horizonCmd.Execute(); err != nil {
-		fmt.Println(err)
-	}
-
-	output := writer.(*bytes.Buffer).String()
-	assert.NotContains(t, output, "--exp-enable-ingestion-filtering")
 }
 
 // Pattern taken from testify issue:
